+++ src/main/java/com/remote_vitals/vital/entities/BloodPressureSystolic.java
@@ -39,7 +39,7 @@ public class BloodPressureSystolic extends VitalRecord {
     }
 
     /******************* Constructors *******************/
-    public BloodPressureSystolic(Long id, float value) {
+    public BloodPressureSystolic(Integer id, float value) {
         super(id);
         setValue(value);
     }
